packages/agents/context-agent/requirements.txt
# Core dependencies for context-agent
websockets>=12.0
aiohttp>=3.9.0
pydantic>=2.0
requests>=2.31.0

# For Ollama API
# No additional dependencies needed

# Logging and monitoring
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.0

# Image metadata handling
Pillow>=10.0.0
piexif>=1.1.3
